feat: init app hook
