+++ src/hooks/useInitApp.tsx
@@ -1,0 +1,81 @@
+import { useEffect, useState } from "react"
+import { Api } from "@/api/base"
+import { CurrentAppResp } from "@/page/App/resp/currentAppResp"
+import { configActions } from "@/redux/config/configSlice"
+import { appInfoActions } from "@/redux/currentApp/appInfo/appInfoSlice"
+import { componentsActions } from "@/redux/currentApp/editor/components/componentsSlice"
+import { actionActions } from "@/redux/currentApp/action/actionSlice"
+import { dragShadowActions } from "@/redux/currentApp/editor/dragShadow/dragShadowSlice"
+import { dottedLineSquareActions } from "@/redux/currentApp/editor/dottedLineSquare/dottedLineSquareSlice"
+import { displayNameActions } from "@/redux/currentApp/displayName/displayNameSlice"
+import { dependenciesActions } from "@/redux/currentApp/executionTree/dependencies/dependenciesSlice"
+import { runAction } from "@/page/App/components/Actions/ActionPanel/utils/runAction"
+import { useParams } from "react-router-dom"
+import { useDispatch } from "react-redux"
+
+export const useInitBuilderApp = () => {
+  let { appId, versionId } = useParams()
+  const dispatch = useDispatch()
+
+  const [loadingState, setLoadingState] = useState(true)
+
+  useEffect(() => {
+    const controller = new AbortController()
+    Api.request<CurrentAppResp>(
+      {
+        url: `/apps/${appId}/versions/${versionId}`,
+        method: "GET",
+        signal: controller.signal,
+      },
+      (response) => {
+        dispatch(configActions.updateIllaMode("edit"))
+        dispatch(appInfoActions.updateAppInfoReducer(response.data.appInfo))
+        dispatch(
+          componentsActions.updateComponentReducer(response.data.components),
+        )
+        dispatch(actionActions.updateActionListReducer(response.data.actions))
+
+        dispatch(
+          dragShadowActions.updateDragShadowReducer(
+            response.data.dragShadowState,
+          ),
+        )
+        dispatch(
+          dottedLineSquareActions.updateDottedLineSquareReducer(
+            response.data.dottedLineSquareState,
+          ),
+        )
+        dispatch(
+          displayNameActions.updateDisplayNameReducer(
+            response.data.displayNameState,
+          ),
+        )
+        dispatch(
+          dependenciesActions.setDependenciesReducer(
+            response.data.dependenciesState,
+          ),
+        )
+        const autoRunAction = response.data.actions.filter((item) => {
+          return item.triggerMode === "automate"
+        })
+        autoRunAction.forEach((item) => {
+          runAction(item)
+        })
+
+        if (response.data.actions.length > 0) {
+          dispatch(configActions.changeSelectedAction(response.data.actions[0]))
+        }
+      },
+      (e) => {},
+      (e) => {},
+      (loading) => {
+        setLoadingState(loading)
+      },
+    )
+    return () => {
+      controller.abort()
+    }
+  }, [])
+
+  return loadingState
+}
